move IsPrime to Extension Method and write Tests
