+++ NumberGen.sln
@@ -11,6 +11,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NumberGen.Data", "NumberGen
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NumberGen.Model", "NumberGen.Model\NumberGen.Model.csproj", "{75D32A97-6057-4B2A-A2D0-37C00F75439C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{3E4E4206-69A8-44F5-A169-B83465428A84}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NumberGenUnitTests", "NumberGenUnitTests\NumberGenUnitTests.csproj", "{64873B6E-930F-4B7F-9380-1B990C03707E}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -29,5 +33,12 @@ Global
 		{75D32A97-6057-4B2A-A2D0-37C00F75439C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{75D32A97-6057-4B2A-A2D0-37C00F75439C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{75D32A97-6057-4B2A-A2D0-37C00F75439C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{64873B6E-930F-4B7F-9380-1B990C03707E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{64873B6E-930F-4B7F-9380-1B990C03707E}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{64873B6E-930F-4B7F-9380-1B990C03707E}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{64873B6E-930F-4B7F-9380-1B990C03707E}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{64873B6E-930F-4B7F-9380-1B990C03707E} = {3E4E4206-69A8-44F5-A169-B83465428A84}
 	EndGlobalSection
 EndGlobal
